Split out `BeatmapOnlineLookupQueue` from `BeatmapManager`
